% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdcor.R
\name{newdcor}
\alias{newdcor}
\title{Calculate the the weighted dcor}
\usage{
newdcor(X, Y, w)
}
\arguments{
\item{X}{the firt variable in arbitrary dimension.}

\item{Y}{the second varaiable, which must be multidimensional.#'}

\item{w}{the weight of each dimension of B}
}
\description{
Calculate the new dcor after paying different weight to every dimension of B.
}
\examples{
\dontrun{
X <- rnorm(10)
Y <- matrix(rnorm(10*5),10,5)
W<- runif(5,0,1)
newdcor(X,Y,w=W)
}
}
