% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.group.R, R/select.top.R
\name{select.group}
\alias{select.group}
\alias{select.top}
\title{Select adjacent parts from Y as a group with the stongest connection with X.}
\usage{
select.group(X, Y, k)

select.top(X, Y, k)
}
\arguments{
\item{X}{the firt variable in arbitrary dimension.}

\item{Y}{the second varaiable, which must be multidimensional.}

\item{k}{the size of the group}

\item{k}{the size of the group}

\item{X}{the firt variable in arbitrary dimension.}

\item{Y}{the second varai, which must be multidimensional.}
}
\description{
Select adjacent parts from Y as a group with the stongest connection with X.

Select the top k'th property of Y as a group with the stongest connection with X
}
\section{Functions}{
\itemize{
\item \code{select.group}: adjacent parts from Y as a group with the stongest connection with X, the connection is measured by "dcor()".

\item \code{select.top}: a group of Y which consisting of the top k'th strongest connectivity parts with X.
}}

\examples{
\dontrun{
X <- rnorm(10)
Y <- matrix(rnorm(10*20),10,20)
select.group(X,Y,k=3)
}
\dontrun{
X <- rnorm(10)
Y <- matrix(rnorm(10*20),10,20)
select.top(X,Y,k=3)
}
}
