% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dcor.test.R
\name{Dcor.test}
\alias{Dcor.test}
\title{Independence test using "dcor"}
\usage{
Dcor.test(x, y, R)
}
\arguments{
\item{x}{data of the first sample}

\item{y}{data of the second sample}

\item{R}{number of replicates}
}
\description{
Independence test using "dcor"
}
\section{Functions}{
\itemize{
\item \code{Dcor.test}: "dcor" to perform a statistical test of dependence with a permutation test.
}}

\examples{
\dontrun{
x <- rnorm(10)
y <- matrix(rnorm(10*5),10,5)
Dcor.test(x,y,R=199)
}
}
