% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei.R
\name{wei}
\alias{wei}
\title{Calculate the correlation between A with each dimension of B.}
\usage{
wei(X, Y)
}
\arguments{
\item{X}{the firt variable in arbitrary dimension.}

\item{Y}{the second varai, which must be multidimensional.}
}
\description{
Using the function "cor()" to measure the connectinity of each dimension of B with A.
}
\examples{
\dontrun{
X <- rnorm(10)
Y <- matrix(rnorm(10*5),10,5)
wei(X,Y)
}
}
